generic: OpenBSD process information fixes

+++ libs/ksysguard/processcore/processes_netbsd_p.cpp
@@ -185,7 +185,7 @@ ProcessesLocal::ProcessesLocal() : d(new Private())
 }
 
 long ProcessesLocal::getParentPid(long pid) {
-    long long ppid = -1;
+    long ppid = -1;
     struct kinfo_proc2 *p;
     if(d->readProc(pid, &p, 0))
     {
@@ -247,23 +247,23 @@ bool ProcessesLocal::setNiceness(long pid, int priority) {
 bool ProcessesLocal::setScheduler(long pid, int priorityClass, int priority)
 {
     if(priorityClass == KSysGuard::Process::Other || priorityClass == KSysGuard::Process::Batch)
-	    priority = 0;
+        priority = 0;
     if(pid <= 0) return false; // check the parameters
     struct sched_param params;
     params.sched_priority = priority;
     switch(priorityClass) {
-      case (KSysGuard::Process::Other):
-	    return (sched_setscheduler( pid, SCHED_OTHER, &params) == 0);
-      case (KSysGuard::Process::RoundRobin):
-	    return (sched_setscheduler( pid, SCHED_RR, &params) == 0);
-      case (KSysGuard::Process::Fifo):
-	    return (sched_setscheduler( pid, SCHED_FIFO, &params) == 0);
+        case (KSysGuard::Process::Other):
+            return (sched_setscheduler( pid, SCHED_OTHER, &params) == 0);
+        case (KSysGuard::Process::RoundRobin):
+            return (sched_setscheduler( pid, SCHED_RR, &params) == 0);
+        case (KSysGuard::Process::Fifo):
+            return (sched_setscheduler( pid, SCHED_FIFO, &params) == 0);
 #ifdef SCHED_BATCH
-      case (KSysGuard::Process::Batch):
-	    return (sched_setscheduler( pid, SCHED_BATCH, &params) == 0);
+        case (KSysGuard::Process::Batch):
+            return (sched_setscheduler( pid, SCHED_BATCH, &params) == 0);
 #endif
-      default:
-	    return false;
+        default:
+            return false;
     }
 }
